Generate Brownie interface for Foundry project
It is now possible to call `moonworm generate-brownie` with the
`--foundry` argument.

If it is called this way, the Brownie interface is generated with
reference to a Foundry project.

Rough around the edges, and may not play nice with `--prod`, which is
anyway not working well for deployments (even before the addition of
`--foundry`).

The same scheme we use to remap the build object for `--foundry` should
be generalized across both `--foundry` and `--prod`.